MW-133 Last fixes
